// frontend/src/api/generated/api.ts
export interface BarChartDataPairUiDTO {
    'categoryId'?: string;
    'data'?: BarChartDataUiDTO;
}
export interface BarChartDataUiDTO {
    'labels'?: Array<string>;
    'expensesData'?: Array<number>;
    'budgetData'?: Array<number>;
}
export interface BudgetCreateRequestUiDTO {
    'name'?: string;
    'description'?: string;
    'isActive'?: boolean;
    'categoryId'?: string;
}
export interface BudgetDTO {
    'id'?: string;
    'name'?: string;
    'description'?: string;
    'isActive'?: boolean;
    'period'?: BudgetDTOPeriodEnum;
    'amount'?: number;
    'periodStart'?: string;
    'periodEnd'?: string;
    'category'?: CategorySimplifiedDTO;
}
// ...
export interface CategoryDTO {
    'id'?: string;
    'name'?: string;
    'description'?: string;
    'isActive'?: boolean;
    'weeklyBudget'?: BudgetDTO;
    'monthlyBudget'?: BudgetDTO;
    'yearlyBudget'?: BudgetDTO;
}
export interface CategoryRepDTO {
    'id'?: string;
    'name'?: string;
    'isActive'?: boolean;
}
// ...
export interface DashboardDataResponseUiDTO {
    'expenses'?: Array<ExpenseUiDTO>;
    'weeklyPieChart'?: PieChartDataUiDTO;
    'monthlyPieChart'?: PieChartDataUiDTO;
    'yearlyPieChart'?: PieChartDataUiDTO;
    'categories'?: Array<CategoryRepDTO>;
    'barChartData'?: Array<BarChartDataPairUiDTO>;
}
export interface ExpenseSearchRequestUiDTO {
    'name'?: string;
    'amountFrom'?: number;
    'amountTo'?: number;
    'description'?: string;
    'isPlanned'?: boolean;
    'dateFrom'?: string;
    'dateTo'?: string;
    'categoryIds'?: Array<string>;
}
export interface ExpenseUiDTO {
    'id'?: string;
    'name'?: string;
    'amount'?: number;
    'description'?: string;
    'isPlanned'?: boolean;
    'date'?: string;
    'category'?: CategoryDTO;
}
// ...
export interface PieChartDataUiDTO {
    'labels'?: Array<string>;
    'data'?: Array<number>;
}